requirements.txt
pybluez
PyYAML
requests
